requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
openai>=1.0.0
requests>=2.31.0
websockets>=11.0.0
mattermostdriver>=7.3.0
python-dotenv>=1.0.0 